# api/openapi.yaml
openapi: 3.0.3
info:
  title: device-cert-issuer API
  version: 1.0.0
  description: Online issuance service for device certificates signed by an Intermediate CA
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/device-certificates:issue:
    post:
      operationId: PostApiV1DeviceCertificatesIssue
      summary: Issue a device certificate from a CSR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueDeviceCertificateRequest"
      responses:
        "200":
          description: Certificate issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDeviceCertificateResponse"
        "400":
          description: Invalid request or CSR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Enrollment or policy rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/server-certificates:issue:
    post:
      operationId: PostApiV1ServerCertificatesIssue
      summary: Issue a server certificate from a CSR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueServerCertificateRequest"
      responses:
        "200":
          description: Server certificate issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueServerCertificateResponse"
        "400":
          description: Invalid request or CSR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Policy rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    IssueDeviceCertificateRequest:
      type: object
      required:
        - device_id
        - tenant_id
        - model
        - csr_pem
      properties:
        device_id:
          type: string
          minLength: 1
          maxLength: 128
        tenant_id:
          type: string
          minLength: 1
          maxLength: 128
        model:
          type: string
          minLength: 1
          maxLength: 128
        csr_pem:
          type: string
          minLength: 1
          maxLength: 16384
        profile:
          type: string
          maxLength: 64
    IssueDeviceCertificateResponse:
      type: object
      required:
        - device_certificate_pem
        - certificate_chain_pem
        - serial_number
        - not_before
        - not_after
      properties:
        device_certificate_pem:
          type: string
        certificate_chain_pem:
          type: string
        serial_number:
          type: string
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
    IssueServerCertificateRequest:
      type: object
      required:
        - csr_pem
      properties:
        csr_pem:
          type: string
          minLength: 1
          maxLength: 16384
    IssueServerCertificateResponse:
      type: object
      required:
        - certificate_pem
        - chain_pem
        - fullchain_pem
        - serial_number
        - not_before
        - not_after
      properties:
        certificate_pem:
          type: string
        chain_pem:
          type: string
        fullchain_pem:
          type: string
        serial_number:
          type: string
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
        subject_summary:
          type: string
        san_summary:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
        request_id:
          type: string
...