# internal/gen/openapi/openapi.gen.go
//go:build go1.22

// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package openapi

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code      string                 `json:"code"`
	Details   *ErrorResponse_Details `json:"details,omitempty"`
	Message   string                 `json:"message"`
	RequestId *string                `json:"request_id,omitempty"`
}

// ErrorResponseDetails0 defines model for .
type ErrorResponseDetails0 map[string]interface{}

// ErrorResponseDetails1 defines model for .
type ErrorResponseDetails1 = []map[string]interface{}

// ErrorResponse_Details defines model for ErrorResponse.Details.
type ErrorResponse_Details struct {
	union json.RawMessage
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// IssueDeviceCertificateRequest defines model for IssueDeviceCertificateRequest.
type IssueDeviceCertificateRequest struct {
	CsrPem   string  `json:"csr_pem"`
	DeviceId string  `json:"device_id"`
	Model    string  `json:"model"`
	Profile  *string `json:"profile,omitempty"`
	TenantId string  `json:"tenant_id"`
}

// IssueDeviceCertificateResponse defines model for IssueDeviceCertificateResponse.
type IssueDeviceCertificateResponse struct {
	CertificateChainPem  string    `json:"certificate_chain_pem"`
	DeviceCertificatePem string    `json:"device_certificate_pem"`
	NotAfter             time.Time `json:"not_after"`
	NotBefore            time.Time `json:"not_before"`
	SerialNumber         string    `json:"serial_number"`
}

// IssueServerCertificateRequest defines model for IssueServerCertificateRequest.
type IssueServerCertificateRequest struct {
	CsrPem string `json:"csr_pem"`
}

// IssueServerCertificateResponse defines model for IssueServerCertificateResponse.
type IssueServerCertificateResponse struct {
	CertificatePem string    `json:"certificate_pem"`
	ChainPem       string    `json:"chain_pem"`
	FullchainPem   string    `json:"fullchain_pem"`
	NotAfter       time.Time `json:"not_after"`
	NotBefore      time.Time `json:"not_before"`
	SanSummary     *string   `json:"san_summary,omitempty"`
	SerialNumber   string    `json:"serial_number"`
	SubjectSummary *string   `json:"subject_summary,omitempty"`
}

// ReadinessResponse defines model for ReadinessResponse.
type ReadinessResponse struct {
	Status string `json:"status"`
}

// PostApiV1DeviceCertificatesIssueJSONRequestBody defines body for PostApiV1DeviceCertificatesIssue for application/json ContentType.
type PostApiV1DeviceCertificatesIssueJSONRequestBody = IssueDeviceCertificateRequest

// PostApiV1ServerCertificatesIssueJSONRequestBody defines body for PostApiV1ServerCertificatesIssue for application/json ContentType.
type PostApiV1ServerCertificatesIssueJSONRequestBody = IssueServerCertificateRequest

// AsErrorResponseDetails0 returns the union data inside the ErrorResponse_Details as a ErrorResponseDetails0
func (t ErrorResponse_Details) AsErrorResponseDetails0() (ErrorResponseDetails0, error) {
	var body ErrorResponseDetails0
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromErrorResponseDetails0 overwrites any union data inside the ErrorResponse_Details as the provided ErrorResponseDetails0
func (t *ErrorResponse_Details) FromErrorResponseDetails0(v ErrorResponseDetails0) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeErrorResponseDetails0 performs a merge with any union data inside the ErrorResponse_Details, using the provided ErrorResponseDetails0
func (t *ErrorResponse_Details) MergeErrorResponseDetails0(v ErrorResponseDetails0) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsErrorResponseDetails1 returns the union data inside the ErrorResponse_Details as a ErrorResponseDetails1
func (t ErrorResponse_Details) AsErrorResponseDetails1() (ErrorResponseDetails1, error) {
	var body ErrorResponseDetails1
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromErrorResponseDetails1 overwrites any union data inside the ErrorResponse_Details as the provided ErrorResponseDetails1
func (t *ErrorResponse_Details) FromErrorResponseDetails1(v ErrorResponseDetails1) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeErrorResponseDetails1 performs a merge with any union data inside the ErrorResponse_Details, using the provided ErrorResponseDetails1
func (t *ErrorResponse_Details) MergeErrorResponseDetails1(v ErrorResponseDetails1) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t ErrorResponse_Details) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *ErrorResponse_Details) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Issue a device certificate from a CSR
	// (POST /api/v1/device-certificates:issue)
	PostApiV1DeviceCertificatesIssue(w http.ResponseWriter, r *http.Request)
	// Issue a server certificate from a CSR
	// (POST /api/v1/server-certificates:issue)
	PostApiV1ServerCertificatesIssue(w http.ResponseWriter, r *http.Request)
	// Liveness probe
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// Readiness probe
	// (GET /readyz)
	GetReadyz(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostApiV1DeviceCertificatesIssue operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1DeviceCertificatesIssue(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostApiV1DeviceCertificatesIssue(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostApiV1ServerCertificatesIssue operation middleware
func (siw *ServerInterfaceWrapper) PostApiV1ServerCertificatesIssue(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostApiV1ServerCertificatesIssue(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReadyz operation middleware
func (siw *ServerInterfaceWrapper) GetReadyz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReadyz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/api/v1/device-certificates:issue", wrapper.PostApiV1DeviceCertificatesIssue)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/server-certificates:issue", wrapper.PostApiV1ServerCertificatesIssue)
	m.HandleFunc("GET "+options.BaseURL+"/healthz", wrapper.GetHealthz)
	m.HandleFunc("GET "+options.BaseURL+"/readyz", wrapper.GetReadyz)

	return m
}

type PostApiV1DeviceCertificatesIssueRequestObject struct {
	Body *PostApiV1DeviceCertificatesIssueJSONRequestBody
}

type PostApiV1DeviceCertificatesIssueResponseObject interface {
	VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error
}

type PostApiV1DeviceCertificatesIssue200JSONResponse IssueDeviceCertificateResponse

func (response PostApiV1DeviceCertificatesIssue200JSONResponse) VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1DeviceCertificatesIssue400JSONResponse ErrorResponse

func (response PostApiV1DeviceCertificatesIssue400JSONResponse) VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1DeviceCertificatesIssue403JSONResponse ErrorResponse

func (response PostApiV1DeviceCertificatesIssue403JSONResponse) VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1DeviceCertificatesIssue500JSONResponse ErrorResponse

func (response PostApiV1DeviceCertificatesIssue500JSONResponse) VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1DeviceCertificatesIssue503JSONResponse ErrorResponse

func (response PostApiV1DeviceCertificatesIssue503JSONResponse) VisitPostApiV1DeviceCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1ServerCertificatesIssueRequestObject struct {
	Body *PostApiV1ServerCertificatesIssueJSONRequestBody
}

type PostApiV1ServerCertificatesIssueResponseObject interface {
	VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error
}

type PostApiV1ServerCertificatesIssue200JSONResponse IssueServerCertificateResponse

func (response PostApiV1ServerCertificatesIssue200JSONResponse) VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1ServerCertificatesIssue400JSONResponse ErrorResponse

func (response PostApiV1ServerCertificatesIssue400JSONResponse) VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1ServerCertificatesIssue403JSONResponse ErrorResponse

func (response PostApiV1ServerCertificatesIssue403JSONResponse) VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1ServerCertificatesIssue500JSONResponse ErrorResponse

func (response PostApiV1ServerCertificatesIssue500JSONResponse) VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PostApiV1ServerCertificatesIssue503JSONResponse ErrorResponse

func (response PostApiV1ServerCertificatesIssue503JSONResponse) VisitPostApiV1ServerCertificatesIssueResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse HealthResponse

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetReadyzRequestObject struct {
}

type GetReadyzResponseObject interface {
	VisitGetReadyzResponse(w http.ResponseWriter) error
}

type GetReadyz200JSONResponse ReadinessResponse

func (response GetReadyz200JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetReadyz503JSONResponse ErrorResponse

func (response GetReadyz503JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Issue a device certificate from a CSR
	// (POST /api/v1/device-certificates:issue)
	PostApiV1DeviceCertificatesIssue(ctx context.Context, request PostApiV1DeviceCertificatesIssueRequestObject) (PostApiV1DeviceCertificatesIssueResponseObject, error)
	// Issue a server certificate from a CSR
	// (POST /api/v1/server-certificates:issue)
	PostApiV1ServerCertificatesIssue(ctx context.Context, request PostApiV1ServerCertificatesIssueRequestObject) (PostApiV1ServerCertificatesIssueResponseObject, error)
	// Liveness probe
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
	// Readiness probe
	// (GET /readyz)
	GetReadyz(ctx context.Context, request GetReadyzRequestObject) (GetReadyzResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// PostApiV1DeviceCertificatesIssue operation middleware
func (sh *strictHandler) PostApiV1DeviceCertificatesIssue(w http.ResponseWriter, r *http.Request) {
	var request PostApiV1DeviceCertificatesIssueRequestObject

	var body PostApiV1DeviceCertificatesIssueJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostApiV1DeviceCertificatesIssue(ctx, request.(PostApiV1DeviceCertificatesIssueRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostApiV1DeviceCertificatesIssue")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostApiV1DeviceCertificatesIssueResponseObject); ok {
		if err := validResponse.VisitPostApiV1DeviceCertificatesIssueResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostApiV1ServerCertificatesIssue operation middleware
func (sh *strictHandler) PostApiV1ServerCertificatesIssue(w http.ResponseWriter, r *http.Request) {
	var request PostApiV1ServerCertificatesIssueRequestObject

	var body PostApiV1ServerCertificatesIssueJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostApiV1ServerCertificatesIssue(ctx, request.(PostApiV1ServerCertificatesIssueRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostApiV1ServerCertificatesIssue")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostApiV1ServerCertificatesIssueResponseObject); ok {
		if err := validResponse.VisitPostApiV1ServerCertificatesIssueResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(w http.ResponseWriter, r *http.Request) {
	var request GetHealthzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetReadyz operation middleware
func (sh *strictHandler) GetReadyz(w http.ResponseWriter, r *http.Request) {
	var request GetReadyzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetReadyz(ctx, request.(GetReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReadyz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetReadyzResponseObject); ok {
		if err := validResponse.VisitGetReadyzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}